backend/requirements.txt
# Backend dependencies for AI-powered contract analysis system
# Target runtime: Python 3.11

# Web Framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.41
alembic==1.16.4
psycopg2-binary
redis==5.2.1

# AI and LLM
openai==1.99.2
anthropic==0.69.0
langchain==0.3.27
langchain-openai==0.3.28
langchain-anthropic==0.3.17
langgraph==0.2.74
llama-index==0.10.68

# Vector Databases
pinecone-client==5.0.1
weaviate-client==4.16.4

# Document Processing
pymupdf==1.26.3
unstructured
python-docx==1.1.2
pytesseract==0.3.13
pillow==10.4.0

# ML and Data
numpy
pandas
scikit-learn
xgboost
shap

# Task Queue
celery==5.5.3
flower==2.0.1

# Utilities
python-dotenv==1.1.1
pydantic==2.12.0
pydantic-settings==2.10.1
httpx==0.28.1
aiofiles==24.1.0
loguru==0.7.3

# Testing
pytest==8.4.1
pytest-asyncio==0.23.6
